+++ saleor/static/dashboard-next/translations/components/TranslationFields/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./TranslationFields";
+export * from "./TranslationFields";
+export * from "./TranslationFieldsSave";