+++ CHANGELOG.md
@@ -35,6 +35,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Display avatars in staff views - #3922 by @dominik-zeglen
 - Prevent user from changing his own status and permissions - #3922 by @dominik-zeglen
 - Support for digital product - #3868 by @korycins
+- Add translation interface - #3884 by @dominik-zeglen
 
 
 ## 2.4.0